BackEnd/requirements.txt
fastapi
uvicorn
httpx
cloudinary
sqlalchemy>=2.0.0
pymysql
fastapi-users[sqlalchemy]>=10.0.0
passlib[bcrypt]
python-jose[cryptography]
python-multipart
email-validator
python-dotenv
asyncmy
PyPDF2
google-genai
# PostgreSQL support in Render
psycopg2-binary
asyncpg>=0.27.0
# SQLite async support
aiosqlite
# Added for SQLAlchemy async support
greenlet>=2.0.0
# Added for better debugging
pytest-asyncio